+++ src/types/modules/thumbs.d.ts
@@ -25,7 +25,7 @@ export interface ThumbsOptions {
    *
    * @default null
    */
-  swiper?: Swiper | null;
+  swiper?: Swiper | string | null;
   /**
    * Additional class that will be added to activated thumbs swiper slide
    *